--- web-common/src/features/tables/config.ts
@@ -1,1 +1,0 @@
-export const OLAP_DRIVERS_WITHOUT_MODELING = ["clickhouse", "druid"];
