+++ web-local/src/routes/(application)/[type=workspace]/[name]/+page.svelte
@@ -327,12 +327,7 @@
           {#if type === "source" && $allErrors[0]?.message}
             <ErrorPane {filePath} errorMessage={$allErrors[0].message} />
           {:else if tableName}
-            <ConnectedPreviewTable
-              {connector}
-              database={undefined}
-              databaseSchema={undefined}
-              table={tableName}
-            />
+            <ConnectedPreviewTable {connector} table={tableName} />
           {/if}
           <svelte:fragment slot="error">
             {#if type === "model"}
